backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==11.0.0
deep_translator==1.11.4
